+++ GrokkingAlgorithms.Lib.Tests/SortHelperTests.cs
@@ -1,11 +1,11 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-using GrokkingAlgorithms.Helpers;
+using GrokkingAlgorithms.Lib;
 using NUnit.Framework;
 using System.Diagnostics;
 	
-namespace GrokkingAlgorithms.Tests.Helpers
+namespace GrokkingAlgorithms.Lib.Tests
 {
 	[TestFixture]
 	public class SortHelperTests